Consolidate generators and refactor to be registered/unregistered as add_removable events
